// src/app/types.d.ts
import { type TODO_FILTERS } from './consts'

export interface Todo {
	id: string
	title: string
	comments: string
	date: string
	completed: boolean
}

export type TodoId = Pick<Todo, 'id'>
export type TodoTitle = Pick<Todo, 'title'>
export type TodoCompleted = Pick<Todo, 'completed '>
export type TodoComments = Pick<Todo, 'comments'>
export type TodoDate = Pick<Todo, 'date'>

export type ListOfTodos = Todo[]

export type FilterValue = typeof TODO_FILTERS[keyof typeof TODO_FILTERS]